+++ packages/gateway/src/errors.js
@@ -10,3 +10,16 @@ export class InvalidUrlError extends Error {
   }
 }
 InvalidUrlError.CODE = 'ERROR_INVALID_URL'
+
+export class TimeoutError extends Error {
+  /**
+   * @param {string} message
+   */
+  constructor(message = 'timeout error') {
+    super(message)
+    this.name = 'TimeoutError'
+    this.status = 408
+    this.code = TimeoutError.CODE
+  }
+}
+TimeoutError.CODE = 'ERROR_TIMEOUT'